Add 'configure', etc., from flex-2.5.4a public release.
The 'configure, 'conf.in', 'install.sh', and 'mkinstalldirs' files
come from:

  http://gnuwin32.sourceforge.net/downlinks/flex-src-zip.php.

which was what I could find as a flex-2.5.4 public release.
It downloads as flex-2.5.4a-1-src.zip and has a sha256sum of
d62ff86fa7ac6219136ea1c23c7c1e32d844ec54bd65dea7bc359d3560e4f7f7.

The 'initscan.c' file is slightly different since the comments
in it have to match what is in 'flex.skl', so it is basically
just what was required to get "make check" to pass.  (But the
only difference from the above zip file is in comments.)