Remove flex.1.
Makefile.in: Remove everything associated with man pages.

Instead, include flex.md in the distribution tarball.  However, it
is not installed by "make install" because I was unable to find any
conventional place it should be put.

Also, in 'dist2', remove all the '@' prefixes so all actions will
be printed.

This completes the process of excising flex.1.