Makefile.in, configure.in, conf.in, parse.y: Remove all 'alloca' stuff.
configure: Regenerate with autoconf.

All of that was apparently due to issues with bison generating a
parser that used alloca but not portably.  Modern bison seems to
(1) not use alloca by default, and (2) when it does, do a better
job of doing so portably.  So flex does not need to deal with it.