conf.in, main.c: Remove NEED_ARGV_FIXUP stuff.
It was added in 06e6f990a9, which references VMS, and I certainly do
not care about VMS support in the short term.  Longer term, this could
be re-added by using an #ifdef inside main.c that specifically checked
for VMS, but I am not going to attempt that now because I would have no
way of testing it.

With this, conf.in and config.h are now empty, so I can get rid of
them entirely in the next commit.