flex.md: Change some *text* to `text`.  Move punctuation after either.
After the pandoc conversion, there was a haphazard mixture of *text*
meaning emphasis, and *text* meaning a piece of syntax.  I changed
a bunch of the latter to `text` but probably missed some due to my
regex being too restrictive.

I'm also changing *text,* to *text*, wherever I see it.