flex.slk: Fix bug: Straighten out the 'isatty' declarations.
There was an #include of <unistd.h>, but that creates portability
issues, and is unnecessary due to the 'isatty' declaration.  It
might have been there because the 'isatty' declaration meant for
use with C++ was in the wrong place.  This would cause compilation
to fail when using a C++ compiler and the C scanner interface when
#include <unistd.h> was removed.

The new 'isatty' properly is declared only if needed based on
the YY_XXX_INTERACTIVE macros, and properly uses extern "C" when
using a C++ compiler.