Add parse.c and parse.h, the outputs from bison-3.0.4.
This is to avoid users needing to have bison for a normal
build from a source distribution.

configure.in: Do not probe for yacc/bison.

Makefile.in: Do not rebuild parse.c unless RUN_BISON=1.
Disable built-in suffix rules.
Include parse.c and parse.h in the source distribution.
Do not use @YACC@.

.gitignore: Stop ignoring parse.c and parse.h.

configure: Regenerate.