configure.in: Reject if srcdir is not ".".
This will not work because Makefile does not use $(srcdir).  That
in turn is mainly because ./configure creates config.mk rather
than Makefile, so running "cd obj; ../configure" would not create
a Makefile in 'obj'.

I could perhaps be persuaded to re-add support for this, but it is
something I never do and only see others doing rarely, so seems not
worth the time and complexity to test and support.