nfa.c: Fix bug: #line directives emitted for continued actions.
Symptom: smflex would emit a #line directive pointing to the source
file for a "continued" action, i.e., one where the action is '|'.
This is wrong because no text from the source file is emitted for
such an action.

Origin: This bug is present in flex-2.5.4.

Fix: In nfa.c, finish_rule(), suppress the line directive for
continued actions.

Add test/vert-bar-action.*.  This test did not fail before because
it does not automatically check line directives, that in turn because
there are still #line problems to be fixed.

With this fix, however, input-scan.lex.c goes from having 2 problems
to just 1.