gen.c, etc.: Fix bug: Insufficient #line directives in block of action text.
Symptom: smflex was emitting the block of user actions with only #line
directives pointing at the source file, even though that block contains
generated text interspersed.  One effect is, when single-stepping in a
debugger, that control seems to flow erratically in the source file,
which in reality it is logically flowing through generated code.

Another symptom is that check-line-directives.py complains at the end
of the action block, again due to generated code being improperly
labeled as source.

Origin: flex-2.5.4 has this bug.

Fix: At the appropriate points, emit #line directives that refer to
the output file so all generated code is labeled accordingly.  The
actual mechanics of the fix are somewhat ugly; see the explanations
added to gen.c.