generated-scanner.skl: Fix bug: Wrong response to yyin->fail().
Symptom: For a C++-interface scanner, if an input error occurs, then
depending on other factors it might be detected by the call to
yyin->fail() in LexerInput().  But the response to that was to return 0,
meaning EOF, instead of -1, meaning error.

Origin: This bug is present in flex-2.5.4.

Fix: Return -1 instead.