input-scan.lex: Fix bug: Wrong #line directive after <<EOF>> continued action.
Symptom: If the scanner source file contains a rule with pattern "<<EOF>>"
and action "|" (a "continued" action), it would emit a bad #line directive.

Origin: This bug is present in flex-2.5.4.

Fix: When we see the "|", emit one more #line directive to cancel the
previous one, as explained in the comments added.