input-scan.lex: Fix bug: Wrong #line directives in section 1 due to blank lines.
Symptom: The #line directives in the generated .c file were wrong in
section 1 if it contained any blank lines.

Origin: This bug is present in flex-2.5.4.

Fix: Remove the "^{OPTWS}{NL}" rule.  That rule could also be fixed by
adding ACTION_ECHO (to echo the blank line into the output), but that
makes it identical to the following rule's action, which works just as
well.

input-scan.lex.c: Regenerate.