main.h, etc.: Rename 'fullspd' to 'jacobson'.
The name "fullspd" is too easy to confuse with "fulltbl".  It also
conveys a goal rather than a method.  The comment in main.h says it
uses the "Jacobson method", from which I took the new name, which
is distinct and memorable.

There is a Wikipedia article on "Succinct data structure" that
references a 1998 PhD thesis by Guy Jacobson and a FOCS 1989 paper
by the same author, so I think that is what is meant here.