generated-header.skl: Rename 'yy_state_type' to 'yy_state_type_t'.
This finishes the job of adding "_t" to the types in the C API.

It is a bit redundant to have something with suffix "_type_t", but
the name "yy_state_t" would be too short IMO, and also would be a
prefix of several other important names, so I prefer this for its
uniqueness.  If I understood a bit more about the algorithm internals
I might be able to choose a name that was similarly unique while also
being more descriptive.