generated-scanner.skl: Fix bug: Memory leak when using start state stack.
Symptom: A scanner that pushes anything onto the state stack would
experience a memory leak because the stack was never freed.

Origin: This bug is present in flex-2.5.4, although in that version,
it arguably only affects the C++ API because the C API only allows
one scanner instance.

Fix: Free the stack when cleaning up.