generated-scanner.skl: Merge all 'yy_is_interactive' into one simple version.
The old code did different things for each of the three ways to
read input.  The line-at-a-time fread case was the most elaborate,
but assumed that newlines were command terminators, and even in
that case was (IMO) doing too much.  Now, we just read one at a
time and it's fine.