smflex.html: Credit for -CJ goes to Van Jacobson, not Guy Jacobson.
I made a mistake in my original inference.  I saw a comment in main.h
(originally in flexdef.h) referring to "Jacobson", searched for that
name, found a Guy Jacobson known for data structure work in the right
time period, and assumed that was who was meant.

However, I later noticed that the Author section explicitly credits
Van Jacobson, who is known for networking algorithms, but was at LBL
in the right time period.  So that's who the comment referred to.