generated-scanner.skl: Rename yy_buf_size to yy_buf_alloc_size.
I want it clear that this is (closely related to!) the allocated
size, not the number of meaningful elements.

Also rename YY_BUF_SIZE to YY_DEFAULT_BUF_ALLOC_SIZE, making its
role clearer, and avoiding having two things with the same name
but different purpose and only differing in capitalization.