generated-scanner.skl: Optimize yy_restart.
When there was no current buffer, the old code would create a new
buffer around the old input stream, then use init_buffer to change
it to the new stream.  Instead, we just create the new buffer around
the new stream to begin with.

Also, the old code called load_buffer twice, once directly and once
inside init_buffer.  The latter was needed because of a flaw in
init_buffer, which set the input stream too late for load_buffer to
see it.  That is now fixed, so yy_restart does not need the extra
load_buffer after init_buffer.