smflex.html: Completely reorganize section structure.
In this pass, I rearranged the sections and their relationships to
form a more coherent, logical structure.  Within the sections
themselves there are only occasional, minor edits, as needed to
ensure links work and text agrees with section titles.  You can
see basically everything that changed just in the TOC changes;
the changes below that are all direct consequences.

Revising the section prose is next.