test/error-handling-cpp.lexpp: Fix bug: Order of eval & print unspecified
In this test, I set a debug flag that causes allocations to print.
The call to e.what() invokes the allocator.  If that is evaluated
before the first '<<', then I see the allocator debug print before
"error: ", but it can also happen in the other order.

The fix is to compute what() in a separate statement.

Same problem and fix for test/large-token-cpp.lexpp.